+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Command `migrate` is introduced for db migration see `Migrate the databases` section of README
 - Command `startdb` starts the cbdb and pcdb database containers only
 - Databases are not deleted after boot2docker restart
+- Import ULU_HOST_ADDRESS and ULU_SULTANS_ADDRESS from Profile
 
 ## [v0.1.0] - 2015-04-16